requirements.txt
alembic==1.12.0
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
PyYAML==6.0.1
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
